requirements.txt
django>=5.2,<5.3
gunicorn>=23,<24
dj-database-url>=3,<4
whitenoise>=6,<7
requests
djangorestframework
tabula-py
pandas

# Uncomment these lines to use a Postgres database. Both are needed, since in production
# (which uses Linux) we want to install from source, so that security updates from the
# underlying Heroku stack image are picked up automatically, thanks to dynamic linking.
# On other platforms/in development, the precompiled binary package is used instead, to
# speed up installation and avoid errors from missing libraries/headers.
#psycopg[c]; sys_platform == "linux"
#psycopg[binary]; sys_platform != "linux"
